Fix to empty the refresh Channel in case the reload method gets error. Some logger fixes.
fix to return the error received from cattle to UI while logging in

Fix to remove the token cookie set by the SAML library.
Fix to passback the saml data in the JWT and verfiy the JWT when it is
received back. No more insecure passing of saml JSON
Fix to not override the SAML provider until reload happens