Reload from updateConfig should not be skipped
It has been seen that the reload call from updateConfig is getting
skipped due to reloads called by cattle on setting changes. This can
lead the provider not initialized with all settings in the config.
So now we make sure that reload after UpdateConfig is done is always processed.
Also save the updateTimeStamp at the end.