Add additional handler to set token as cookie
RedirectBackBase might differ from the acs URL. In this case, we
will create a token in the ACS handler and sent it as a hidden field
in a form with html and js to submit the form on load, and this will
be sent to the redirectBackBase URL with new handler for creating token
cookie. This new handler will receive token and set it as a cookie

RedirectBackBase & RedirecBackPath won't be set on the global
SamlServiceProvider struct. HandleSamlLogin sets relayState per request,
which contains the URI of the request, which has the redirectBackBase and
RedirectBackPath values. This commit retrieves those values in the
getPossibleRequestIDs method, along with the requestID, and passes along
that to HandleSamlAssertion. So we don't need to set any values separately.

The setting to specify whitelisted domains for redirectBackBase values will
accept domain. So the users can enter domain.com or domain.com:8080, this
commit checks the redirectBackBase against both.